local kvstore working for backend state tests, new test not checkpointed somehow
